Added OnUpdate hook
